+++ 3rdparty/exported/quickjs/quickjs-time.h
@@ -2,11 +2,19 @@
 
 int qjs_gettimeofday(struct timeval* tv, void* tz)
 {
+  if (tv != NULL)
+  {
+    memset(tv, 0, sizeof(struct timeval));
+  }
   return 0;
 }
 
 struct tm* qjs_localtime_r(const time_t* timep, struct tm* result)
 {
+  if (result != NULL)
+  {
+    memset(result, 0, sizeof(struct tm));
+  }
   return 0;
 }