+++ src/node/node_state.h
@@ -239,7 +239,7 @@ namespace ccf
       const std::vector<uint8_t>& expected_node_public_key_der,
       CodeDigest& code_digest) override
     {
-      return EnclaveAttestationProvider::verify_quote_against_store(
+      return AttestationProvider::verify_quote_against_store(
         tx, quote_info, expected_node_public_key_der, code_digest);
     }
 
@@ -282,7 +282,7 @@ namespace ccf
     //
     void launch_node()
     {
-      auto code_id = EnclaveAttestationProvider::get_code_id(quote_info);
+      auto code_id = AttestationProvider::get_code_id(quote_info);
       if (code_id.has_value())
       {
         node_code_id = code_id.value();
@@ -336,8 +336,8 @@ namespace ccf
     {
       auto fetch_endorsements =
         [this](
-          QuoteInfo quote_info_,
-          const pal::EndorsementEndpointConfiguration& config) {
+          const QuoteInfo& quote_info_,
+          const pal::snp::EndorsementEndpointsConfiguration& endpoint_config) {
           if (quote_info_.format != QuoteFormat::amd_sev_snp_v1)
           {
             // Note: Node lock is already taken here as this is called back
@@ -351,11 +351,10 @@ namespace ccf
             return;
           }
 
-          quote_endorsements_client =
-            std::make_shared<QuoteEndorsementsClient>(rpcsessions);
-
-          quote_endorsements_client->fetch_endorsements(
-            config, [this, quote_info_](std::vector<uint8_t>&& endorsements) {
+          quote_endorsements_client = std::make_shared<QuoteEndorsementsClient>(
+            rpcsessions,
+            endpoint_config,
+            [this, quote_info_](std::vector<uint8_t>&& endorsements) {
               // Note: Only called for SEV-SNP
               std::lock_guard<pal::Mutex> guard(lock);
               quote_info = quote_info_;
@@ -363,6 +362,8 @@ namespace ccf
               launch_node();
               quote_endorsements_client.reset();
             });
+
+          quote_endorsements_client->fetch_endorsements();
         };
 
       pal::attestation_report_data report_data = {};
@@ -371,7 +372,10 @@ namespace ccf
         node_pub_key_hash.h.begin(),
         node_pub_key_hash.h.end(),
         report_data.begin());
-      pal::generate_quote(report_data, fetch_endorsements);
+      pal::generate_quote(
+        report_data,
+        fetch_endorsements,
+        config.attestation.snp_endorsements_servers);
     }
 
     NodeCreateInfo create(StartType start_type_, StartupConfig&& config_)
@@ -2316,7 +2320,7 @@ namespace ccf
         *node_sign_kp,
         sig_tx_interval,
         sig_ms_interval,
-        false /* start time signatures after first tx */);
+        false /* start timed signatures after first tx */);
       network.tables->set_history(history);
     }