+++ tests/infra/remote.py
@@ -597,6 +597,7 @@ class CCFRemote(object):
         node_data_json_file=None,
         service_cert_file="service_cert.pem",
         service_data_json_file=None,
+        snp_endorsements_servers=None,
         **kwargs,
     ):
         """
@@ -667,6 +668,21 @@ class CCFRemote(object):
                 "challenge_server_interface"
             ] = host.acme_challenge_server_interface
 
+        # SNP endorsements servers
+        snp_endorsements_servers = snp_endorsements_servers or []
+        snp_endorsements_servers_list = []
+        for s in snp_endorsements_servers:
+            try:
+                server_type, url = s.split(":")
+            except ValueError as e:
+                raise ValueError(
+                    "SNP endorsements servers should be in the format type:url"
+                ) from e
+            s = {}
+            s["type"] = server_type
+            s["url"] = url
+            snp_endorsements_servers_list.append(s)
+
         # Configuration file
         if config_file:
             LOG.info(
@@ -704,6 +720,7 @@ class CCFRemote(object):
                 node_data_json_file=node_data_json_file,
                 service_data_json_file=service_data_json_file,
                 service_cert_file=service_cert_file,
+                snp_endorsements_servers=snp_endorsements_servers_list,
                 **kwargs,
             )